Change: Add Unify and Bind trait to separate infer implementation
