+++ src/transform/type_infer/type_infer.rs
@@ -17,7 +17,7 @@ use crate::transform::Transform;
 
 use super::poly_type::PolyType;
 use super::subst::Subst;
-use super::traits::Types;
+use super::traits::{Types, Unify};
 use super::type_::TypeVarGen;
 use super::type_env::TypeEnv;
